Move aggregates on meaningful locs to locations
